requirements.txt
boto3==1.35.72
celery==5.4.0
cfenv==0.5.3
cryptography==44.0.0
dj_database_url==2.3.0
django-cors-headers==4.6.0
django-deprecate-fields==0.2.1
django-migration-linter==5.1.0
django-structlog==9.0.0
Django==5.1.5
djangorestframework==3.15.2
drf-spectacular==0.27.2
git+https://github.com/fecgov/fecfile-validate@736ccb5530c1edda426ffac12ad6bcf3b9ba9423#egg=fecfile_validate&subdirectory=fecfile_validate_python
github3.py==4.0.1
GitPython==3.1.43
gunicorn==23.0.0
invoke==2.2.0
jwcrypto==1.5.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==4.5.5
requests==2.32.3
structlog==24.4.0
zeep==4.2.1
